--- a/sorovi.DependencyInjection.AutoRegister.Benchmark/Program.cs
+++ b/sorovi.DependencyInjection.AutoRegister.Benchmark/Program.cs
@@ -1,0 +1,7 @@
+﻿// See https://aka.ms/new-console-template for more information
+
+using BenchmarkDotNet.Running;
+using sorovi.DependencyInjection.AutoRegister.Benchmark;
+
+var summary = BenchmarkRunner.Run<AddingToServiceCollectionBenchmark>();
+