--- a/sorovi.DependencyInjection.AutoRegister.Benchmark/sorovi.DependencyInjection.AutoRegister.Benchmark.csproj
+++ b/sorovi.DependencyInjection.AutoRegister.Benchmark/sorovi.DependencyInjection.AutoRegister.Benchmark.csproj
@@ -1,0 +1,19 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+    <PropertyGroup>
+        <OutputType>Exe</OutputType>
+        <TargetFramework>net6.0</TargetFramework>
+        <ImplicitUsings>enable</ImplicitUsings>
+        <Nullable>enable</Nullable>
+    </PropertyGroup>
+
+    <ItemGroup>
+      <PackageReference Include="BenchmarkDotNet" Version="0.13.2" />
+    </ItemGroup>
+
+    <ItemGroup>
+      <ProjectReference Include="..\sorovi.DependencyInjection.AutoRegister.TestAssembly\sorovi.DependencyInjection.AutoRegister.TestAssembly.csproj" />
+      <ProjectReference Include="..\sorovi.DependencyInjection.AutoRegister\sorovi.DependencyInjection.AutoRegister.csproj" />
+    </ItemGroup>
+
+</Project>