--- a/sorovi.DependencyInjection.AutoRegister.sln
+++ b/sorovi.DependencyInjection.AutoRegister.sln
@@ -9,6 +9,8 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "sorovi.DependencyInjection.
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "sorovi.DependencyInjection.AutoRegister.TestAssembly", "sorovi.DependencyInjection.AutoRegister.TestAssembly\sorovi.DependencyInjection.AutoRegister.TestAssembly.csproj", "{3903ECF7-4D2B-45F0-B0ED-1EF275A2EBCD}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "sorovi.DependencyInjection.AutoRegister.Benchmark", "sorovi.DependencyInjection.AutoRegister.Benchmark\sorovi.DependencyInjection.AutoRegister.Benchmark.csproj", "{26449A80-05F3-4C6C-8135-0AECAC605990}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -58,5 +60,17 @@ Global
 		{3903ECF7-4D2B-45F0-B0ED-1EF275A2EBCD}.Release|x64.Build.0 = Release|Any CPU
 		{3903ECF7-4D2B-45F0-B0ED-1EF275A2EBCD}.Release|x86.ActiveCfg = Release|Any CPU
 		{3903ECF7-4D2B-45F0-B0ED-1EF275A2EBCD}.Release|x86.Build.0 = Release|Any CPU
+		{26449A80-05F3-4C6C-8135-0AECAC605990}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{26449A80-05F3-4C6C-8135-0AECAC605990}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{26449A80-05F3-4C6C-8135-0AECAC605990}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{26449A80-05F3-4C6C-8135-0AECAC605990}.Debug|x64.Build.0 = Debug|Any CPU
+		{26449A80-05F3-4C6C-8135-0AECAC605990}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{26449A80-05F3-4C6C-8135-0AECAC605990}.Debug|x86.Build.0 = Debug|Any CPU
+		{26449A80-05F3-4C6C-8135-0AECAC605990}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{26449A80-05F3-4C6C-8135-0AECAC605990}.Release|Any CPU.Build.0 = Release|Any CPU
+		{26449A80-05F3-4C6C-8135-0AECAC605990}.Release|x64.ActiveCfg = Release|Any CPU
+		{26449A80-05F3-4C6C-8135-0AECAC605990}.Release|x64.Build.0 = Release|Any CPU
+		{26449A80-05F3-4C6C-8135-0AECAC605990}.Release|x86.ActiveCfg = Release|Any CPU
+		{26449A80-05F3-4C6C-8135-0AECAC605990}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 EndGlobal