--- a/sorovi.DependencyInjection.AutoRegister/ServiceCollectionExtension.cs
+++ b/sorovi.DependencyInjection.AutoRegister/ServiceCollectionExtension.cs
@@ -62,18 +62,11 @@ namespace sorovi.DependencyInjection.AutoRegister
 
             var types = assemblies
                 .SelectMany(assembly => assembly.GetExportedTypes())
-                .Where(type =>
-                {
-                    var defaultCondition = type.IsClass && !type.IsAbstract && !type.IsNested && !type.IsGenericType && type.CustomAttributes.Any();
-                    if (predicate is null) { return defaultCondition; }
-
-                    return defaultCondition && predicate(type);
-                });
+                .Where(type => IsOfInterest(type) && HasRegisterAttributes(type) && (predicate is null || predicate(type)));
 
             foreach (var type in types)
             {
                 var attribute = type.GetCustomAttribute(typeof(ServiceAttribute)) ?? type.GetCustomAttribute(typeof(BackgroundServiceAttribute));
-                if (attribute is null) { continue; }
 
                 switch (attribute)
                 {
@@ -88,10 +81,7 @@ namespace sorovi.DependencyInjection.AutoRegister
                             _ => throw new Exception($"unknown lifetime attribute: {attribute.GetType().FullName}")
                         };
 
-                        var serviceDescriptor = serviceAttribute.InterfaceType is null
-                            ? ServiceDescriptor.Describe(type, type, serviceLifetime)
-                            : ServiceDescriptor.Describe(serviceAttribute.InterfaceType, type, serviceLifetime);
-
+                        var serviceDescriptor = ServiceDescriptor.Describe(serviceAttribute.InterfaceType ?? type, type, serviceLifetime);
                         AddServiceDescriptor(services, serviceAttribute.Mode, serviceDescriptor);
                         break;
 
@@ -107,7 +97,11 @@ namespace sorovi.DependencyInjection.AutoRegister
             services.AddSingleton(new AlreadyKnownAssemblies(alreadyKnownAssemblies.KnownAssemblies.Concat(assemblies).ToArray()));
         }
 
-        private static void AddServiceDescriptor(IServiceCollection services, Mode mode, ServiceDescriptor descriptor)
+        private static bool HasRegisterAttributes(in MemberInfo type) => type.CustomAttributes.Any() && type.GetCustomAttributes().Any(attr => attr is ServiceAttribute or BackgroundServiceAttribute);
+
+        private static bool IsOfInterest(in Type type) => type.IsClass && !type.IsAbstract && !type.IsGenericType && !type.IsNested;
+
+        private static void AddServiceDescriptor(in IServiceCollection services, in Mode mode, in ServiceDescriptor descriptor)
         {
             switch (mode)
             {
@@ -132,7 +126,7 @@ namespace sorovi.DependencyInjection.AutoRegister
                 KnownAssemblies = Array.Empty<Assembly>();
             }
 
-            public AlreadyKnownAssemblies(Assembly[] knownAssemblies)
+            public AlreadyKnownAssemblies(in Assembly[] knownAssemblies)
             {
                 KnownAssemblies = knownAssemblies;
             }