… README.md
… sorovi.DependencyInjection.AutoRegister/sorovi.DependencyInjection.AutoRegister.csproj