# Before

|       Method |        Mean |    Error |   StdDev |  Ratio | RatioSD |   Gen0 | Allocated | Alloc Ratio |
|------------- |------------:|---------:|---------:|-------:|--------:|-------:|----------:|------------:|
|  ManuallyAdd |    236.7 ns |  1.31 ns |  1.23 ns |   1.00 |    0.00 | 0.1030 |     648 B |        1.00 |
| AutoRegister | 28,514.1 ns | 93.21 ns | 77.83 ns | 120.58 |    0.71 | 0.9460 |    6020 B |        9.29 |