@andreas-soroko 2022-12-17
# Current

|       Method |       Mean |   Error |  StdDev | Ratio | RatioSD |   Gen0 | Allocated | Alloc Ratio |
|------------- |-----------:|--------:|--------:|------:|--------:|-------:|----------:|------------:|
|  ManuallyAdd |   235.6 ns | 0.77 ns | 0.72 ns |  1.00 |    0.00 | 0.1030 |     648 B |        1.00 |
| AutoRegister | 1,786.5 ns | 5.23 ns | 4.64 ns |  7.58 |    0.03 | 0.1297 |     824 B |        1.27 |