@andreas-soroko 2022-12-17
# Current

|       Method |        Mean |     Error |    StdDev |  Ratio | RatioSD |   Gen0 | Allocated | Alloc Ratio |
|------------- |------------:|----------:|----------:|-------:|--------:|-------:|----------:|------------:|
|  ManuallyAdd |    238.5 ns |   0.88 ns |   0.83 ns |   1.00 |    0.00 | 0.1030 |     648 B |        1.00 |
| AutoRegister | 26,079.9 ns | 287.45 ns | 240.03 ns | 109.29 |    0.98 | 0.6104 |    3979 B |        6.14 |