@andreas-soroko 2022-12-17
# Current

|       Method |       Mean |   Error   |  StdDev  | Ratio | RatioSD |   Gen0 | Allocated | Alloc Ratio |
|------------- |------------:|---------:|---------:|------:|--------:|-------:|----------:|------------:|
|  ManuallyAdd |    240.9 ns |  4.58 ns |  4.70 ns |  1.00 |    0.00 | 0.1032 |    648 B  |        1.00 |
| AutoRegister | 20,649.6 ns | 85.00 ns | 75.35 ns | 85.51 |    1.54 | 0.2441 |   1617 B  |        2.50 |